Commit swagger documentation
Having swagger installed with npm wasn't going to work
because we needed to make substantial changes to the package's
default index page for our purposes. I committed swagger to the
project since we'll want the swagger docs to always be
available with the project without extra effort and we want
the default json file to be our projects and not the sample
petstore.