Add manifest for user-testing app
We want an app that can be deployed independently from the main
line branch so that we can make quick changes and have a public URL
available for user testing different interfaces we make over the
next few sprints.

This can be deployed by anybody with:
cf push -f manifest-user-testing.yml

It should be deployed to:
https://crime-data-api-user-testing.fr.cloud.gov

This is not yet set up for continuous deployement (CD) because
it is unknown if we will be using the same branch continually
for testing and because its extra work that doesn't seem
valuable right now.