Add past_10_years_reported to cde_agencies
This is a measure of how many times in the past decade of SRS reports,
the given agency has reported 12 months of data. This is relative to
the most recent year of SRS reports not the current calendar year (so
for instance, right now we only have 2014 data so it's counting the
years 2005-2014).