Remove /counts/ API endpoint, reta_month_offense_subcat_summary
This was a useful API endpoint to have as we developed the
application, but it is rather large as a table and takes forever to
regenerate and is only used by Arson counts anymore (all other state
and county crimes use the Estimated data endpoint). So removing it for
now.