WIP: Fixing a few things around CSV responses
The old PR for adding Surrogate-Control was trying to parse all
responses into JSON which caused it to crash when the user provided
the output=csv. So I decided to rewrite the wrapper for cache control
to use our own wrapper that could take a surrogate caching time as
well. This lets us remove one more dependency from the code.

While I was at it, I decided to change it that all endpoints should
call CdeResource.render_response instead of with_metadata to ensure
that all paths return a Flask Response object. And I added some tests
to look for caching headers for a few of them.

The issue that isn't working right now are the endpoints that use the
database to return JSON. Right now, the results are being escaped
within a "json_data": [...] element and I don't know how to inline
that JSON into the larger paginated JSON so I can avoid parsing it.

Also to be done: add more assertions to tests that we can download
various endpoints as CSV data and that cache headers are correct