Remove some remaining authentication remnants
When they first created this app, they used a Flask skeleton app which
had support for user login and the like. This API has no user
logins (only a single shared token used by the API frontend proxy), so
I removed most of that in PR #121

But there were some references to setting a SECRET in the docs and in
the environment variable, so need to clean those up.

Also, removing http auth which was only there for protecting the
Swagger.json page before we started ATO process.