Initial work on offense-specific offender view changes
The idea is to join against the offenses to get all offender-offense
pairs. To get started, I also loaded in all NIBRS records for RI in
2014 in the test database

Since the NIBRS data is huge, this commit also replaces test_db with
gzipped file instead. I don't know why we didn't think to do this
earlier, but I guess it never was the most urgent change