Add a content security policy (CSP) header
Add a restrictive CSP HTTP header for all requests.
This prevents the browser/CSP-responsive user agent
from loading any resources that are not from the
same domain as the app.

This does break the swagger UI as it loads a number
of third party resources, but we want people to use
the stylized swagger UI within the frontend app
anyway.